Launch acq task with many queues in drift mode
Fixes https://github.com/HiPERCAM/hiperCamCCD/issues/7 by launching the acquisition task in drift mode with the correct number of queues! 

Note that this is a partial fix; there should be many fewer dropped frames in drift mode now, but dropped frames or missing timestamps are still invisible to the user. It would be good to add an error message to the acquisition task, but this will require a bit of thought to do it the "right" way